phase 2/pyproject.toml
[project]
name = "graphical-context-extraction-phase2"
version = "0.1.0"
description = "Phase 2: Alignment, Chunking, Embeddings and Storage"
requires-python = ">=3.10"
dependencies = [
    "sentence-transformers>=2.2.0",
    "open-clip-torch>=2.20.0",
    "qdrant-client>=1.7.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "scikit-learn>=1.3.0",
    "Pillow>=10.0.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "numpy>=1.24.0",
    "transformers>=4.30.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]
